Move TopBar component and related styles to its own directory
